+++ addon/components/mdc-form-field.js
@@ -13,4 +13,11 @@ export default Component.extend({
   classNames: ['mdc-form-field'],
   classNameBindings: ['align-end:mdc-form-field--align-end'],
   //endregion
+
+  //region Methods
+  createFoundation() {
+    // TODO: Mix in MDCComponent and implement an adapter for MDCFormFieldAdapter
+    // https://material.io/develop/web/components/input-controls/form-fields/#usage-within-web-frameworks
+  },
+  //endregion
 });