+++ addon/components/mdc-list/divider.js
@@ -4,6 +4,12 @@ import layout from '../../templates/components/mdc-list/divider';
 export default Component.extend({
   //region Attributes
   /**
+   * Leaves gaps on each side of divider to match padding of {{mdc-list/item/meta}}
+   * @type {Boolean}
+   */
+  padded: false,
+  /**
+   * Increases the leading margin of the divider so that it does not intersect the avatar column
    * @type {Boolean}
    */
   inset: false,
@@ -13,7 +19,7 @@ export default Component.extend({
   layout,
   tagName: 'li',
   classNames: ['mdc-list-divider'],
-  classNameBindings: ['inset:mdc-list-divider--inset'],
+  classNameBindings: ['padded:mdc-list-divider--padded', 'inset:mdc-list-divider--inset'],
   attributeBindings: ['role'],
   //endregion