+++ addon/components/mdc-list/item.js
@@ -1,17 +1,33 @@
 import Component from '@ember/component';
+import { get, computed } from '@ember/object';
 import events from '../../utils/global-event-handlers';
 import layout from '../../templates/components/mdc-list/item';
 
 export default Component.extend({
   //region Attributes
   /**
+   * Multiple items can be selected at the same time when using the selected state.
+   *
+   * Selected state should be used when it is likely to change soon
+   * (e.g. selecting one or more photos to share in Google Photos).
+   *
+   * @type {Boolean}
+   */
+  selected: false,
+  /**
+   * Activated state is similar to selected state, however
+   * should only be used once within a specific list.
+   *
+   * Activated state is more permanent than selected state,
+   * and will NOT change soon relative to the lifetime of the page.
+   *
    * @type {Boolean}
    */
   activated: false,
   /**
    * @type {Boolean}
    */
-  selected: false,
+  'two-line': false,
   //endregion
 
   //region Ember Hooks
@@ -19,6 +35,12 @@ export default Component.extend({
   tagName: 'li',
   classNames: ['mdc-list-item'],
   classNameBindings: ['activated:mdc-list-item--activated', 'selected:mdc-list-item--selected'],
-  attributeBindings: [...events],
+  attributeBindings: ['aria-selected', ...events],
+  //endregion
+
+  //region Computed Properties
+  'aria-selected': computed('selected', function() {
+    return get(this, 'selected') ? 'true' : null;
+  }),
   //endregion
 });